Implement a non-runtime-dependent workaround for non-seeded init
This patch implements some heavy reworking of the Mersenne Twister
implementation.  In particular, the heavy lifting of both the `seed`
methods and `popFront` have been farmed out to static implementation
methods not dependent on `this`, while the generator's internal state
has been reworked as a dedicated struct.

Taken together these changes allow for the creation of another static
method, `defaultState`, which can be used to default-initialize RNG
state via CTFE.  This default will be entirely overwritten whenever the
RNG is explicitly seeded, but avoids the need for runtime `if` checks
should the generator be default-initialized.